Fixing missing hhsuite_tsv_file.ff{data,index}
Bc of the bracktes chars some shells (like mine) could not be able to do str
expansion so i chance it to be two string hhsuite_tsv_file.ffdata anda hhsuite_tsv_file.ffindex.
I also change the call from a os.run call for a subprocess call with the flag check=True
this way the return code of the the index subprocess will be captured and bugs on this spet can be
detected.